+++ build_test/lib/build_test.dart
@@ -12,6 +12,7 @@ export 'src/in_memory_writer.dart';
 export 'src/matchers.dart';
 export 'src/multi_asset_reader.dart' show MultiAssetReader;
 export 'src/package_reader.dart' show PackageAssetReader;
+export 'src/record_logs.dart';
 export 'src/resolve_source.dart';
 export 'src/stub_reader.dart';
 export 'src/stub_writer.dart';